moves setSearchQuery and resetSearchQuery to the containers/App
